refactor: move the messageEnum from types to enum
