+++ src/controllers/project.controller.ts
@@ -2,6 +2,7 @@ import { Deployment } from "@ak-proto-ts/deployments/v1/deployment_pb";
 import { Project } from "@ak-proto-ts/projects/v1/project_pb";
 import { Session } from "@ak-proto-ts/sessions/v1/session_pb";
 import { RequestHandler } from "@controllers/utilities/requestHandler";
+import { MessageType } from "@enums";
 import { translate } from "@i18n";
 import { IProjectView } from "@interfaces";
 import {
@@ -10,7 +11,6 @@ import {
 	ProjectsService,
 	SessionsService,
 } from "@services";
-import { MessageType } from "@type";
 import { getIds } from "@utilities/getIds.utils";
 import { MessageHandler } from "@views";
 import isEqual from "lodash/isEqual";
