+++ src/enums/index.ts
@@ -1,1 +1,2 @@
-export { Theme } from "@enums/theme.enums";
+export { Theme } from "@enums/theme.enum";
+export { MessageType } from "@enums/messageType.enum";
