+++ src/enums/messageType.enum.ts
@@ -1,0 +1,8 @@
+export enum MessageType {
+	setDeployments = "SET_DEPLOYMENTS",
+	setProjectName = "SET_PROJECT_NAME",
+	setTheme = "SET_THEME",
+	runProject = "RUN_PROJECT",
+	buildProject = "BUILD_PROJECT",
+	setSessions = "SET_SESSIONS",
+}
