+++ src/types/index.ts
@@ -1,4 +1,4 @@
 export { HttpHeaders, RequestConfig } from "@type/api.type";
-export { Message, MessageType } from "@type/message.type";
+export { Message } from "@type/message.type";
 export { LocalhostConnection } from "@type/connection.type";
 export { ApplyResponse } from "@type/entities";
