+++ tsconfig.json
@@ -10,7 +10,6 @@
 			"@ak-proto-ts/*": ["autokitteh/proto/gen/ts/autokitteh/*"],
 			"@api": ["api/index"],
 			"@api/*": ["api/*"],
-			"@connection": ["connection"],
 			"@constants": ["constants/index"],
 			"@constants/*": ["constants/*"],
 			"@controllers": ["controllers/index"],
@@ -19,6 +18,8 @@
 			"@enums/*": ["enums/*"],
 			"@i18n/*": ["i18n/*"],
 			"@i18n": ["i18n/index"],
+			"@models": ["models/index"],
+			"@models/*": ["models/*"],
 			"@interfaces": ["interfaces/index"],
 			"@interfaces/*": ["interfaces/*"],
 			"@providers": ["providers/index"],
